requirements.txt
python-dotenv
flask-sqlalchemy
flask-migrate
flask-script
flask-restplus
pytest
PyMySQL
requests
pika
redis